+++ app/views/are_we_there_yet/splash.rb
@@ -11,17 +11,21 @@ class AreWeThereYet
 
       def open
         sync_exec {
-          @splash_image = image(File.expand_path(File.join('..', '..', '..', '..', 'are-we-there-yet-logo.png'), __FILE__)).scale_to(128, 128)
+          @splash_image = image(File.expand_path(File.join('..', '..', '..', '..', 'are-we-there-yet-splash.gif'), __FILE__)).scale_to(118, 118)
           @shell_proxy = shell(:no_trim, :on_top) {
             minimum_size 128, 128
-            background rgb(33, 44, 186)
+#             background rgb(33, 44, 186)
+            background :white
             grid_layout(1, false) {
               margin_width 0
               margin_height 0
-            }            
-            label {
-              background :transparent
-              image @splash_image
+            }
+            composite {
+              layout_data(:center, :center, true, true) {              
+                width_hint 118
+                height_hint 118
+              }
+              background_image @splash_image              
             }
             cursor display.swt_display.get_system_cursor(swt(:cursor_appstarting))
           }
@@ -30,7 +34,7 @@ class AreWeThereYet
       end
 
       def close
-        sync_exec {
+        async_exec {
           @shell_proxy&.swt_widget&.close
         }
       end